move utility code into pkg/util
Continuing to clean up the main directory, this patch moves utility code
like Warn(), SafeCloser() etc into a `ghw/pkg/util` package.
